packages/python/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "boto3>=1.39.5",
    "docker>=7.1.0",
    "litestar>=2.19.0",
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
    "botocore>=1.42.24",
    "starlette>=0.50.0",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = ["boto3-stubs[batch,ecs,secretsmanager,s3,lambda]>=1.40.29"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
force-include = { "src/common/py.typed" = "common/py.typed" }
